move another fixture
